Move generic query params dto to improve folders organization
