+++ src/common/base/application/dto/query-params/get-all-options.interface.ts
