+++ src/common/base/application/service/crud-service.interface.ts
@@ -1,6 +1,6 @@
 import { CollectionDto } from '@common/base/application/dto/collection.dto';
 import { IDto, IResponseDto } from '@common/base/application/dto/dto.interface';
-import { IGetAllOptions } from '@common/base/application/dto/get-all-options.interface';
+import { IGetAllOptions } from '@common/base/application/dto/query-params/get-all-options.interface';
 import { SuccessOperationResponseDto } from '@common/base/application/dto/success-operation-response.dto';
 import IEntity from '@common/base/domain/entity.interface';
