+++ src/module/category/application/dto/query-params/category-fields-query-params.dto.ts
@@ -1,7 +1,7 @@
 import { Transform } from 'class-transformer';
 import { IsIn, IsOptional } from 'class-validator';
 
-import { IGetAllOptions } from '@common/base/application/dto/get-all-options.interface';
+import { IGetAllOptions } from '@common/base/application/dto/query-params/get-all-options.interface';
 import { fromCommaSeparatedToArray } from '@common/base/application/mapper/base.mapper';
 
 import { Category } from '@module/category/domain/category.entity';
