+++ src/module/payment-method/application/dto/query-params/payment-method-fields-query-params.dto.ts
@@ -1,7 +1,7 @@
 import { Transform } from 'class-transformer';
 import { IsIn, IsOptional } from 'class-validator';
 
-import { IGetAllOptions } from '@common/base/application/dto/get-all-options.interface';
+import { IGetAllOptions } from '@common/base/application/dto/query-params/get-all-options.interface';
 import { fromCommaSeparatedToArray } from '@common/base/application/mapper/base.mapper';
 
 import { PaymentMethod } from '@module/payment-method/domain/payment-method.entity';
