+++ src/module/payment-method/interface/payment-method.controller.ts
@@ -13,7 +13,7 @@ import {
 
 import { Hypermedia } from '@common/base/application/decorator/hypermedia.decorator';
 import { CollectionDto } from '@common/base/application/dto/collection.dto';
-import { PageQueryParamsDto } from '@common/base/application/dto/page-query-params';
+import { PageQueryParamsDto } from '@common/base/application/dto/query-params/page-query-params';
 import { SuccessOperationResponseDto } from '@common/base/application/dto/success-operation-response.dto';
 import { HttpMethod } from '@common/base/application/enum/http-method.enum';
